lab_4/lab4/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["datacurrentsea <giomola@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10 <3.11"
uvicorn = "^0.20.0"
requests = "^2.28.2"
pytest = "^7.2.1"
httpx = "^0.23.3"
isort = "^5.12.0"
black = "^22.12.0"
fastapi = "^0.89.1"
scipy = "^1.10"
scikit-learn = "1.2.1"
coverage = "^7.2.2"
fastapi-cache2 = "0.1.9"
redis = "^4.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
